apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-ai-swarm-api"
version = "0.1.0"
description = "FastAPI backend for the Autonomous AI Swarm platform"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "beautifulsoup4>=4.12.3",
  "email-validator>=2.2.0",
  "fastapi>=0.115.6",
  "httpx>=0.28.1",
  "minio>=7.2.15",
  "openai>=1.58.1",
  "openpyxl>=3.1.5",
  "orjson>=3.10.12",
  "pandas>=2.2.3",
  "playwright>=1.49.1",
  "pgvector>=0.3.6",
  "psycopg[binary]>=3.2.3",
  "pydantic-settings>=2.7.0",
  "pypdf>=5.1.0",
  "python-slugify>=8.0.4",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.36",
  "sqlmodel>=0.0.22",
  "sse-starlette>=2.2.1",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "faker>=33.1.0",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.0",
  "ruff>=0.8.4"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
